refactor: Address first comments in 1st delivery
- Change the way to write the Routes
- Change the icons folder location
- Add the useEffect to use YouTube API
