+++ src/pages/Video/mock.js
@@ -1,0 +1,9 @@
+const mockVideo = {
+  id: 'uE-1RPDqJAY',
+  snippet: {
+    description: 'THEY"RE TAKING THE HOBBITS TO ISENGARD (not my work)',
+    title: 'taking the hobbits to isengard',
+  },
+};
+
+export default mockVideo;
